requirements.txt
comtypes==1.2.0
Flask==3.0.0
Flask_Cors==4.0.0
mediapipe==0.10.7
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
psutil==5.9.5
pycaw==20230407
pydub==0.25.1
python-dotenv==1.0.0
Requests==2.31.0
screen_brightness_control==0.22.1
